added pagination with knp Paginator small reworks

+++ src/Controller/JobControllers/JobDelController.php
@@ -23,6 +23,10 @@ class JobDelController extends AbstractController{
     public function projectsCreatePage(int $id, Request $request): Response
     {
         $job = $this->jobRepository->find($id);
+        if ($job === null) {
+            return $this->redirectToRoute('company_list_jobs', [
+            ], Response::HTTP_MOVED_PERMANENTLY);
+        }
         $deleteJob =$this->jobRepository->deleteJob();
         $delJob = $this->createForm(JobType::class, $job);
         $delJob->handleRequest($request);
